Extract RebindToMachinePredicate from JcloudsLocation
